A new ActivityList created, implemented a new Homepage
